# code/backend/metiche/rest/server/openapi.yaml
...
openapi: 3.1.0
info:
    title: Metiche
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /accounts/{id}:
        get:
            tags:
                - Account
            summary: Get Account by ID
            operationId: getAccount
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Account'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - Account
            summary: Partially update Account
            operationId: patchAccount
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Account'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Account'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - Account
            summary: Delete Account by ID
            operationId: deleteAccount
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /invites/{id}:
        get:
            tags:
                - Invite
            summary: Get Invite by ID
            operationId: getInvite
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Invite'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - Invite
            summary: Partially update Invite
            operationId: patchInvite
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Invite'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Invite'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - Invite
            summary: Delete Invite by ID
            operationId: deleteInvite
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /limit-events/{id}:
        get:
            tags:
                - LimitEvent
            summary: Get LimitEvent by ID
            operationId: getLimitEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LimitEvent'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - LimitEvent
            summary: Partially update LimitEvent
            operationId: patchLimitEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LimitEvent'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LimitEvent'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - LimitEvent
            summary: Delete LimitEvent by ID
            operationId: deleteLimitEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /notification-channels/{id}:
        get:
            tags:
                - NotificationChannel
            summary: Get NotificationChannel by ID
            operationId: getNotificationChannel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotificationChannel'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - NotificationChannel
            summary: Partially update NotificationChannel
            operationId: patchNotificationChannel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NotificationChannel'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotificationChannel'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - NotificationChannel
            summary: Delete NotificationChannel by ID
            operationId: deleteNotificationChannel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
        Account:
            properties:
                claimed_at:
                    format: date-time
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                display_name:
                    type: string
                email:
                    format: email
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                identity_handle:
                    nullable: true
                    type: string
                identity_provider:
                    $ref: '#/components/schemas/IdentityProvider'
                identity_subject:
                    nullable: true
                    type: string
                key:
                    type: string
                last_seen_at:
                    format: date-time
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/RecordStatus'
                token_hash:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - key
                - display_name
                - token_hash
                - identity_provider
                - status
                - created_at
                - updated_at
            type: object
...
        ConflictSeverity:
            enum:
                - invalid
                - low
                - medium
                - high
                - critical
            type: string
...
        DeliveryStatus:
            enum:
                - invalid
                - never_sent
                - ok
                - failing
                - disabled_after_failures
            type: string
...
        IdentityProvider:
            enum:
                - invalid
                - none
                - github
            type: string
...
        Invite:
            properties:
                code:
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by_member_uuid:
                    format: uuid
                    nullable: true
                    type: string
                expires_at:
                    format: date-time
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                label:
                    nullable: true
                    type: string
                last_used_at:
                    format: date-time
                    nullable: true
                    type: string
                max_uses:
                    format: int64
                    nullable: true
                    type: integer
                revoked_at:
                    format: date-time
                    nullable: true
                    type: string
                revoked_by_member_uuid:
                    format: uuid
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/InviteStatus'
                team_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
                uses:
                    format: int64
                    type: integer
            required:
                - id
                - team_uuid
                - code
                - uses
                - status
                - created_at
                - updated_at
            type: object
...
        InviteStatus:
            enum:
                - invalid
                - active
                - exhausted
                - expired
                - revoked
            type: string
...
        LimitEvent:
            properties:
                allowed:
                    format: int64
                    nullable: true
                    type: integer
                created_at:
                    format: date-time
                    type: string
                dimension:
                    type: string
                id:
                    format: uuid
                    type: string
                note:
                    nullable: true
                    type: string
                observed:
                    format: int64
                    type: integer
                occurred_at:
                    format: date-time
                    type: string
                outcome:
                    $ref: '#/components/schemas/LimitOutcome'
                plan_uuid:
                    format: uuid
                    nullable: true
                    type: string
                session_uuid:
                    format: uuid
                    nullable: true
                    type: string
                team_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - team_uuid
                - dimension
                - observed
                - outcome
                - occurred_at
                - created_at
                - updated_at
            type: object
...
        LimitOutcome:
            enum:
                - invalid
                - warned
                - refused
                - allowed_over
            type: string
...
        NotificationChannel:
            properties:
                consecutive_failures:
                    format: int64
                    type: integer
                created_at:
                    format: date-time
                    type: string
                created_by_member_uuid:
                    format: uuid
                    nullable: true
                    type: string
                delivery_status:
                    $ref: '#/components/schemas/DeliveryStatus'
                id:
                    format: uuid
                    type: string
                key:
                    type: string
                kind:
                    $ref: '#/components/schemas/NotificationChannelKind'
                label:
                    type: string
                last_attempt_at:
                    format: date-time
                    nullable: true
                    type: string
                last_error:
                    nullable: true
                    type: string
                last_success_at:
                    format: date-time
                    nullable: true
                    type: string
                min_severity:
                    $ref: '#/components/schemas/ConflictSeverity'
                notify_human_requests:
                    type: boolean
                project_uuid:
                    format: uuid
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/RecordStatus'
                target_url:
                    type: string
                team_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - team_uuid
                - key
                - kind
                - label
                - target_url
                - min_severity
                - notify_human_requests
                - status
                - delivery_status
                - consecutive_failures
                - created_at
                - updated_at
            type: object
        NotificationChannelKind:
            enum:
                - invalid
                - slack
                - discord
                - webhook
            type: string
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object
...
        RecordStatus:
            enum:
                - invalid
                - active
                - inactive
            type: string
...